requirements.txt
boto3
pyyaml
prometheus_client